remove kton module
