+++ Cargo.toml
@@ -56,11 +56,9 @@ members = [
     "node/runtime",
     "node/rpc-client",
     "srml/staking",
-    "srml/kton",
     "srml/support",
     "srml/aura",
     "srml/try",
-    "srml/balances-ext",
     "srml/reward",
 ]
