+++ node/runtime/Cargo.toml
@@ -38,12 +38,10 @@ serde = { version = "1.0", optional = true }
 substrate-keyring = { git = 'https://github.com/paritytech/substrate.git', optional = true }
 # customed
 #ring = { package = "evo-ring", path = '../../srml/token/ring', default-features = false}
-kton = { package = "evo-kton", path = '../../srml/kton', default-features = false}
 staking = { package = "evo-staking", path = "../../srml/staking", default-features = false}
 aura = { package = "srml-aura",  path = "../../srml/aura", default-features = false }
 #aura = { package = "srml-aura", path = "../../srml/aura", default-features = false}
 reward = { package = "evo-reward",  path = "../../srml/reward", default-features = false }
-balances-ext = { package = "evo-balances-ext", path = "../../srml/balances-ext", default-features = false}
 
 [features]
 default = ["std"]
@@ -81,7 +79,5 @@ std = [
 	"rustc-hex",
 	"substrate-keyring",
 	"offchain-primitives/std",
-	"kton/std",
   "reward/std",
-  "balances-ext/std",
 ]
