--- srml/kton/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "evo-kton"
-version = "0.1.0"
-authors = ["Darwinia Network <hello@itering.com>"]
-edition = "2018"
-
-[dependencies]
-serde = { version = "1.0", optional = true }
-safe-mix = { version = "1.0", default-features = false}
-parity-codec = { version = "4.1.1", default-features = false, features = ["derive"] }
-substrate-keyring = { git = 'https://github.com/paritytech/substrate.git', optional = true }
-rstd = { package = "sr-std", git = 'https://github.com/paritytech/substrate.git', default-features = false }
-primitives = { package = "sr-primitives", git = 'https://github.com/paritytech/substrate.git', default-features = false }
-srml-support = { git = 'https://github.com/paritytech/substrate.git', default-features = false }
-system = { package = "srml-system", git = 'https://github.com/paritytech/substrate.git', default-features = false }
-timestamp = { package = "srml-timestamp", git = 'https://github.com/paritytech/substrate.git', default-features = false }
-substrate-primitives = { git = 'https://github.com/paritytech/substrate.git', default-features = false }
-dsupport = { package = "evo-support", path = "../support", default-features = false }
-
-[dev-dependencies]
-runtime_io = { package = "sr-io", git = 'https://github.com/paritytech/substrate.git' }
-substrate-primitives = { git = 'https://github.com/paritytech/substrate.git' }
-balances = { package = "srml-balances", git = 'https://github.com/paritytech/substrate.git' }
-node-runtime = { path = "../../node/runtime" }
-
-[features]
-default = ["std"]
-std = [
-	"serde",
-	"safe-mix/std",
-	"substrate-keyring",
-	"parity-codec/std",
-	"rstd/std",
-	"srml-support/std",
-	"primitives/std",
-	"system/std",
-	"timestamp/std",
-	"substrate-primitives/std",
-	"dsupport/std",
-]
