+++ srml/reward/src/lib.rs
@@ -20,6 +20,35 @@ use system::ensure_signed;
 use dsupport::traits::OnAccountBalanceChanged;
 use dsupport::traits::OnDilution;
 
+const DEPOSIT_ID: LockIdentifier = *b"lockkton";
+
+#[derive(Encode, Decode, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "std", derive(Debug))]
+pub struct BalanceLock<Balance, BlockNumber> {
+    pub id: LockIdentifier,
+    pub amount: Balance,
+    pub until: BlockNumber,
+    pub reasons: WithdrawReasons,
+}
+
+
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Encode, Decode, Default)]
+#[cfg_attr(feature = "std", derive(Debug))]
+pub struct IndividualDeposit<Currency, Balance, Moment> {
+    pub month: u32,
+    pub start_at: Moment,
+    pub value: Currency,
+    pub balance: Balance,
+    pub claimed: bool,
+}
+
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Encode, Decode, Default)]
+#[cfg_attr(feature = "std", derive(Debug))]
+pub struct Deposit<Currency, Balance, Moment> {
+    pub total: Currency,
+    pub deposit_list: Vec<IndividualDeposit<Currency, Balance, Moment>>,
+}
+
 type KtonBalanceOf<T> = <<T as Trait>::Kton as Currency<<T as system::Trait>::AccountId>>::Balance;
 type RingBalanceOf<T> = <<T as Trait>::Ring as Currency<<T as system::Trait>::AccountId>>::Balance;
 
@@ -28,7 +57,7 @@ pub type RingPositiveImbalanceOf<T> = <<T as Trait>::Ring as Currency<<T as syst
 
 pub trait Trait: timestamp::Trait {
     type Kton: Currency<Self::AccountId>;
-    type Ring: Currency<Self::AccountId>;
+    type Ring: LockableCurrency<Self::AccountId, Moment=Self::BlockNumber>;
 
     type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
 }
@@ -36,8 +65,13 @@ pub trait Trait: timestamp::Trait {
 decl_event!(
     pub enum Event<T> where
     < T as system::Trait>::AccountId,
+    Balance = KtonBalanceOf<T>,
     Currency = RingBalanceOf<T>,
+    Moment = < T as timestamp::Trait>::Moment,
     {
+        /// lock ring for getting kton
+        NewDeposit(Moment, AccountId, Balance, Currency),
+
         /// Claim Reward
         RewardClaim(AccountId, Currency),
     }
@@ -45,17 +79,22 @@ decl_event!(
 
 decl_storage! {
     trait Store for Module<T: Trait> as Reward {
-        // reward you can get per kton
-        pub RewardPerShare get(reward_per_share): RingBalanceOf<T>;
-        // reward already paid to each ktoner
-        pub RewardPaidOut get(reward_paid_out): map T::AccountId => i128;
+        pub DepositLedger get(deposit_ledger): map T::AccountId => Option<Deposit<RingBalanceOf<T>, KtonBalanceOf<T>, T::Moment>>;
 
         pub SysAcc get(sys_acc) config(): T::AccountId;
+
+        // reward you can get per kton
+        pub RewardPerShare get(reward_per_share): RingBalanceOf<T>;
+
+        // reward already paid to each ktoner
+        pub RewardPaidOut get(reward_paid_out): map T::AccountId => i128;
 
         /// system revenue
         /// same to balance in ring
         /// TODO: it's ugly, ready for hacking
         pub SysRevenuePot get(system_revenue): map T::AccountId => RingBalanceOf<T>;
+
+
     }
     add_extra_genesis {}
 }
@@ -63,6 +102,53 @@ decl_storage! {
 decl_module! {
     pub struct Module<T: Trait> for enum Call where origin: T::Origin {
         fn deposit_event<T>() = default;
+
+        pub fn deposit(origin, value: RingBalanceOf<T>, months: u32) {
+            ensure!(!months.is_zero() && months <= 36, "months must be at least 1");
+            let transactor = ensure_signed(origin)?;
+            if <DepositLedger<T>>::exists(&transactor) {
+                return Err("Already deposited.");
+            }
+
+            let free_currency = T::Ring::free_balance(&transactor);
+            let value = value.min(free_currency);
+
+            let now = <timestamp::Module<T>>::now();
+
+            let kton_return = Self::compute_kton_balance(months, value).unwrap();
+
+            let individual_deposit = IndividualDeposit {month: months, start_at: now.clone(), value: value, balance: kton_return, claimed: false};
+            let deposit = Deposit {total: value, deposit_list: vec![individual_deposit]};
+
+            Self::update_deposit(&transactor, &deposit);
+
+            let _positive_imbalance = T::Kton::deposit_into_existing(&transactor, kton_return);
+            Self::kton_minted(&transactor, kton_return);
+            Self::deposit_event(RawEvent::NewDeposit(now, transactor, kton_return, value));
+        }
+
+        fn deposit_extra(origin, additional_value: RingBalanceOf<T>, months: u32) {
+            ensure!(!months.is_zero() && months <= 36, "months must be at least 1");
+            let transactor = ensure_signed(origin)?;
+            let mut deposit = Self::deposit_ledger(&transactor).ok_or("Use fn deposit instead.")?;
+
+            let now = <timestamp::Module<T>>::now();
+            let free_currency = T::Ring::free_balance(&transactor);
+
+            if let Some(extra) = free_currency.checked_sub(&deposit.total) {
+                let extra = extra.min(additional_value);
+                deposit.total += extra;
+
+                let kton_return = Self::compute_kton_balance(months, extra).unwrap();
+                let individual_deposit = IndividualDeposit {month: months, start_at: now.clone(), value: extra.clone(), balance: kton_return, claimed: false};
+                deposit.deposit_list.push(individual_deposit);
+                Self::update_deposit(&transactor, &deposit);
+
+                let _positive_imbalance = T::Kton::deposit_into_existing(&transactor, kton_return);
+                Self::kton_minted(&transactor, kton_return);
+                Self::deposit_event(RawEvent::NewDeposit(now, transactor, kton_return, extra));
+            }
+        }
 
         pub fn claim_reward(origin) {
             let transactor = ensure_signed(origin)?;
@@ -76,7 +162,44 @@ decl_module! {
     }
 }
 
+
 impl<T: Trait> Module<T> {
+
+    fn kton_minted(who: &T::AccountId, value: KtonBalanceOf<T>) {
+        let additional_reward_paid_out = Self::convert_to_paid_out(value);
+        Self::update_reward_paid_out(who, additional_reward_paid_out);
+    }
+
+    fn update_deposit(who: &T::AccountId, deposit: &Deposit<RingBalanceOf<T>, KtonBalanceOf<T>, T::Moment>) {
+        T::Ring::set_lock(
+            DEPOSIT_ID,
+            &who,
+            deposit.total,
+            // u32::max_value().into(),
+            T::BlockNumber::max_value(),
+            WithdrawReasons::all()
+        );
+        <DepositLedger<T>>::insert(who, deposit);
+    }
+
+    fn compute_kton_balance(months: u32, value: RingBalanceOf<T>) -> Option<KtonBalanceOf<T>> {
+        let months = months as u64;
+        let value = value.try_into().unwrap_or_default() as u64;
+
+        if !months.is_zero() {
+            let no = U256::from(67_u128).pow(U256::from(months));
+            let de = U256::from(66_u128). pow(U256::from(months));
+
+            let quotient = no / de;
+            let remainder = no % de;
+            let res = U256::from(value) * (U256::from(1000) * (quotient - 1) + U256::from(1000) * remainder / de) / U256::from(1970000);
+
+            Some(res.as_u64().try_into().unwrap_or_default())
+        } else {
+            None
+        }
+
+    }
 
     fn convert_to_paid_out(value: KtonBalanceOf<T>) -> RingBalanceOf<T> {
         let value: u64 = value.try_into().unwrap_or_default() as u64;
@@ -152,15 +275,6 @@ impl<T: Trait> Module<T> {
             Ok((system_imbalance, acc_imbalance))
         }
 
-}
-
-/// account kton balance changed
-impl<T: Trait> OnAccountBalanceChanged<T::AccountId, KtonBalanceOf<T>> for Module<T> {
-    fn on_changed(who: &T::AccountId, old: KtonBalanceOf<T>, new: KtonBalanceOf<T>) {
-        // update reward paid out
-        let additional_reward_paid_out = Self::convert_to_paid_out(new-old);
-        Self::update_reward_paid_out(who, additional_reward_paid_out);
-    }
 }
 
 /// reward(ring minted)
