+++ srml/staking/Cargo.toml
@@ -22,7 +22,6 @@ dsupport = { package = "evo-support", path = "../support", default-features = fa
 substrate-primitives = { git = 'https://github.com/paritytech/substrate.git' }
 timestamp = { package = "srml-timestamp", git = 'https://github.com/paritytech/substrate.git' }
 balances = { package = "srml-balances", git = 'https://github.com/paritytech/substrate.git' }
-kton = { package = "evo-kton", path = "../kton" }
 rand = "0.6.5"
 node-runtime = { path = "../../node/runtime" }
